WIP using models instead of controllers
